In Sep10ServiceTest, pull the tests that connects to ledgerClient to another integration test
