+++ essential-tests/build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
   // Stellar dependencies
   testFixturesImplementation(libs.stellar.wallet.sdk)
   testFixturesImplementation(variantOf(libs.java.stellar.sdk) { classifier("uber") })
+  testFixturesImplementation(libs.okhttp3.mockserver)
 
   // Spring dependencies
   testFixturesImplementation("org.springframework.boot:spring-boot-starter-data-jpa")
